+++ web/src/stream_edit.ts
@@ -261,6 +261,7 @@ export function show_settings_for(node: HTMLElement): void {
             realm.realm_org_type === settings_config.all_org_type_values.business.code,
         is_admin: current_user.is_admin,
         org_level_message_retention_setting: get_display_text_for_realm_message_retention_setting(),
+        realm_mobile_push_notifications_enabled: realm.realm_mobile_push_notifications_enabled,
         can_access_stream_email: stream_data.can_access_stream_email(sub),
         group_setting_labels: settings_config.all_group_setting_labels.stream,
     });